requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# LangChain and related packages
langchain>=0.1.0
langchain-core>=0.1.46
langchain-community>=0.0.16
langchain-openai>=0.0.5
langchain-chroma>=0.0.10

# Vector store and embeddings
chromadb>=0.4.22
openai>=1.12.0
tiktoken>=0.5.2

# PDF processing
pypdf>=3.0.0
